Prepare resume transfers from directory in Conn, implements PeerInfo result, correct closable resources usage in Conn

+++ src/main/java/org/jed2k/Peer.java
@@ -23,7 +23,7 @@ public class Peer implements Comparable<Peer> {
 
     long    lastConnected   = 0;
     long    nextConnection  = 0;
-    long    failCount       = 0;
+    int     failCount       = 0;
     boolean connectable     = false;
     int source      = 0;
     NetworkIdentifier   endpoint;
